pyproject.toml
[project]
name = "mutab"
version = "0.1.0"
dependencies = [
  "apted",
  "distance",
  "local_attention",
  "lxml",
  "mmcv-full<2",
  "mmdet<3",
  "mmocr<1",
  "more-itertools",
  "numpy",
  "positional-encodings[pytorch]",
  "ranger@git+https://github.com/lessw2020/Ranger-Deep-Learning-Optimizer",
  "torch>=2,<2.1",
  "tqdm",
  "urllib3<2",
  "yapf==0.40.1",
]

[tool.setuptools.packages.find]
include = ["mutab"]
